requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
flask-sqlalchemy>=3.0.0
flask-dance>=7.0.0
flask-migrate>=4.0.0
flask-mail>=0.10.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-auth>=2.0.0
gunicorn>=21.2.0
PyJWT>=2.8.0
cryptography>=41.0.0
